Revert to older version of pCCDS (seniority free flavor).
Changes:
- Add back generation of power series to describe an N-th order pair
excitation in terms of as all possible pair excitation orders, up to N-th.
This gave a list of dicts with hole/particle indexes for each occupation
vector, from which the permanents were latter computed.
- The overlap derivative that we had was for the pCCDS_sen-o variant of
pCCDS, not the seniority free one, to which we are going back. For now declare
as not implemented.